objAnimFn_80038f38: (int) cast on u8 bank entry ==1 for signed cmpwi (#58/#69) (97.71->98.21)
The entries[entryIdx] == 1 compare on an unsigned u8 emitted cmplwi; target
uses signed cmpwi (matching the sibling (int)entries[idx] != 0xff cast). Cast
to int. Residual is a #108 loop register permutation.

+++ src/main/objprint.c
@@ -82,7 +82,7 @@ void objAnimFn_80038f38(int obj, char* p2)
         {
             u8* entries = *(u8**)&((ObjDef*)m)->jointData;
             int idx = OBJPRINT_ACTIVE_BANK_INDEX(obj) + entryIdx + 1;
-            if ((int)entries[idx] != 0xff && entries[entryIdx] == 1)
+            if ((int)entries[idx] != 0xff && (int)entries[entryIdx] == 1)
             {
                 found = (s16*)((char*)((GameObject*)obj)->anim.jointPoseData + vecOffset);
             }
